requirements.txt
# LLM & Embeddings (Optimized)
langchain>=0.3.0
langchain-community>=0.3.0
ollama>=0.4.0
FlagEmbedding>=1.2.10  # For BGE-M3 multi-functionality (dense+sparse+colbert)
sentence-transformers>=2.2.2  # Required by FlagEmbedding

# Vector Database
qdrant-client>=1.12.0

# Document Processing
pypdf>=5.0.0
python-docx>=1.1.0

# API & Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
requests>=2.32.0
gradio>=4.0.0  # Web Interface

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.8.0
tqdm>=4.66.0

# Caching & Performance
diskcache>=5.6.3
cachetools>=5.3.0

# Better JSON handling
orjson>=3.9.0